removed GrpCarriageIntakeDualControl
Got rid of the command group in favor of just using CarriageIntakeControl command as default command for CarriageIntake default command.  Created static variable for elevatorHeight in Robot.java to store elevator position that can be referenced anywhere.

+++ src/org/usfirst/frc/team226/robot/subsystems/Elevator.java
@@ -108,7 +108,7 @@ public class Elevator extends Subsystem {
 		return left.getClosedLoopError(Constants.ELEVATOR_PID_IDX);
 	}
 	
-	public double getElevatorPosition() {
+	public int getElevatorPosition() {
 		return left.getSelectedSensorPosition(Constants.ELEVATOR_PID_IDX);
 	}
